the program doe snot remove the dash (-) and I want it also able to recognize name of the file. it does not know how to save the new files with the database that you just made with the "submit entered text" form with a database loaded. so each word must contain enough of information that you can access (name of the file in which it was taken) (date) type of file. this can be added to json file depending on the part of the program in which it is.
First= the parsing of text (my program does not recognize such word)

Second= each word must have much dataset so that it can be easier then to manage all words, all files, all users and so on (for example if each word has as much info as it can have then later you can access the word in context.
This can only be done if the database json is very wlel referenced.